% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM_helpers.R
\name{assign_IPCs}
\alias{assign_IPCs}
\title{A set of small functions for restructuring data and running linear models.
Restructures IPCs into column-wise form for easier use in comparisons.}
\usage{
assign_IPCs(
  num_IPC,
  IPC_names,
  All_IPCs,
  test_n,
  train_n,
  testIPCs,
  trainIPCs,
  num_Comparisons
)
}
\value{
A list of two lists, the first containing the restructured test IPCs and the second containing
the restructured train IPCs, ready to be used elsewhere in the analysis.
}
\description{
A set of small functions for restructuring data and running linear models.
Restructures IPCs into column-wise form for easier use in comparisons.
}
