% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_BIC_paths.R
\name{check_levels}
\alias{check_levels}
\title{Runs the loop that checks whether additional paths are needed to find the best IPC fit. This function is run in a loop
by \code{compare_IPC()}.}
\usage{
check_levels(
  checking_levels,
  BICS_Output1,
  Prior_Best_IPCs,
  Do_Check,
  Level_Paths,
  num_IPC,
  IPC_names,
  IPC_prior_level,
  ROI_Train,
  trainIPCs,
  Lv_list,
  DataType2,
  model_improved,
  num_chains,
  All_Paths,
  ThisPath,
  BestBic,
  output_path
)
}
\value{
\code{num_chains}: The number of paths that were searched in order to find the IPC that best fits the data. The lower this
number, the fewer paths it took to reach the best-matching IPC (which may suggest that the IPC contributed more strongly
to stimulus representations in that ROI).
}
\description{
Runs the loop that checks whether additional paths are needed to find the best IPC fit. This function is run in a loop
by \code{compare_IPC()}.
}
