% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_BIC_paths.R
\name{compare_IPC}
\alias{compare_IPC}
\title{A series of functions to compare IPCs and determine which IPCs provide the best fit to the data. All of these
functions are used within the loop of \code{run_BIC()}.
Tracks, identifies, and records the best-matching IPCs, along with associated statistical information, for output.}
\usage{
compare_IPC(
  i,
  num_IPC,
  CurrentLevel,
  IPC_names,
  Best1,
  ROI_Train,
  DataType2,
  loop_end,
  trainIPCs,
  testIPCs,
  Lv_list,
  test_Lv_list,
  BICS_Output1,
  All_Paths,
  ROI,
  Analysis,
  num_chains,
  ThisPath
)
}
\value{
A list of outputs to facilitate the updating of IPCs and associated BIC scores, as follows:
\itemize{
\item \code{loop_end}: Set to 1 if the current best BIC score matches the previous best (i.e. no more improvement in the model);
defaults to 0 during run of loop.
\item \code{Lv_list}: The list of best-matching IPCs for training data for a given ROI, updated every loop. IPCs that have not been
identified in the ROI are assigned a matrix of 1s in this list.
\item \code{test_Lv_list}: Same as \code{Lv_list}, except for test data (especially for use in cross-validation).
\item \code{BICS_Output_1}: The row of BIC scores for each IPCs, for output to the Summary table.
\item \code{Best1}: The index of the best-matching IPC for a given ROI.
\item \code{All_Paths}: A list of the paths that were searched in order to find the IPC that best fits the data.
\item \code{num_chains}: The number of paths that were searched in order to find the IPC that best fits the data. The lower this
number, the fewer paths it took to reach the best-matching IPC (which may suggest that the IPC contributed more strongly
to stimulus representations in that ROI).
\item \code{This_Path}: The current path that was searched in order to find the IPC that best fits the data.
\item \code{BestBic}: The BIC score for the best-matching IPC for a given ROI.
}
}
\description{
A series of functions to compare IPCs and determine which IPCs provide the best fit to the data. All of these
functions are used within the loop of \code{run_BIC()}.
Tracks, identifies, and records the best-matching IPCs, along with associated statistical information, for output.
}
