% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_BIC_paths.R
\name{follow_paths}
\alias{follow_paths}
\title{Checks individual levels of the IPC search path, and determines whether the model has improved or not.}
\usage{
follow_paths(
  checking_level,
  Level2Check,
  Splits,
  BICS_Output1,
  BICS_Output_alternate,
  Prior_Best_IPCs,
  Do_Check,
  Level_Paths,
  num_IPC,
  IPC_names,
  IPC_prior_level,
  ROI_Train,
  trainIPCs,
  Lv_list,
  DataType2,
  model_improved,
  num_chains,
  All_Paths,
  ThisPath,
  BestBic,
  output_path,
  Chain_Header,
  ROI,
  Analysis
)
}
\value{
A list of outputs indicating the results of the level check, as follows:
\itemize{
\item \code{model_improved}: Has a value of 1 if the model moved after the new path search (which continues the loop). If
the model did not improve after the last search, this has a value of 0, which breaks the search loop and outputs
the results of the search to a series of tables.
\item \code{num_chains}: The number of paths that were searched in order to find the IPC that best fits the data. The lower this
number, the fewer paths it took to reach the best-matching IPC (which may suggest that the IPC contributed more strongly
to stimulus representations in that ROI).
}
}
\description{
Checks individual levels of the IPC search path, and determines whether the model has improved or not.
}
