% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_calc_BIC.R
\name{get_data_by_ROI}
\alias{get_data_by_ROI}
\title{A series of functions for extracting data for each ROI and conducting a Bayesian Information Criterion (BIC)
analysis on data from each ROI.
Extracts data for each ROI, then runs Bayesian Information Criterion (BIC) analyses and returns a summary table. This
function is run in a loop by \code{run_PCM()} for each ROI.}
\usage{
get_data_by_ROI(
  ROI_num,
  ROIs,
  Analysis,
  data_file,
  Summary,
  train_n,
  test_n,
  data_name,
  repetitions,
  num_IPC,
  num_Comparisons,
  testIPCs,
  trainIPCs,
  SampleSize,
  input_path,
  output_path
)
}
\value{
A Summary matrix, containing summary information on the best-matching IPCSs for output to the final data file.
}
\description{
A series of functions for extracting data for each ROI and conducting a Bayesian Information Criterion (BIC)
analysis on data from each ROI.
Extracts data for each ROI, then runs Bayesian Information Criterion (BIC) analyses and returns a summary table. This
function is run in a loop by \code{run_PCM()} for each ROI.
}
