% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POIs_to_dfs.R
\name{POIs_to_dfs}
\alias{POIs_to_dfs}
\title{A function that takes individual rows of a dataframe defining all POIs and
converts it into its own dataframe for comparison to input data.}
\usage{
POIs_to_dfs(POI_row)
}
\description{
A function that takes individual rows of a dataframe defining all POIs and
converts it into its own dataframe for comparison to input data.
}
