% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_inputs.R
\name{process_inputs}
\alias{process_inputs}
\title{A function that processes inputs to run a Pattern Component Modelling analysis of Representational Similarity
Analysis (RSA) data.}
\usage{
process_inputs(input_filename)
}
\arguments{
\item{input_filename}{A file containing \emph{representational similarity analysis (RSA)} data for a given \emph{region of interest (ROI)},
in the form \verb{[sample size x number of comparisons]}. This file's name should be of the format \code{ROI_vert.csv}, where \code{ROI} is the
name of a brain region from which the data was obtained (e.g. \code{ACC}).}
}
\description{
A function that processes inputs to run a Pattern Component Modelling analysis of Representational Similarity
Analysis (RSA) data.
}
