% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_calc_BIC.R
\name{run_BIC}
\alias{run_BIC}
\title{Returns Bayesian Information Criterion (BIC) information comparing participant data for that ROI to the input IPCs,
in order to determine which IPCs best fit the data. This function is run in a loop by \code{get_data_by_ROI()}.}
\usage{
run_BIC(
  mci,
  num_IPC,
  IPC_names,
  train_n,
  test_n,
  num_Comparisons,
  ROI_Data,
  CV_Log,
  ROI,
  testIPCs,
  trainIPCs,
  DataType2,
  Analysis,
  ROI_num,
  Summary,
  SampleSize,
  output_path
)
}
\value{
A list of outputs containing information about the IPC fits in the dataset, as follows:
\itemize{
\item \code{Summary}: A summary table of the BIC outputs, to be used in the output tables.
\item \code{Best_Output}: The IPC that best fits the data, according to BIC.
\item \code{All_Paths}: A list of the paths that were searched in order to find the IPC that best fits the data.
\item \code{CV_Log}: If cross-validation was conducted, an output of each iteration of cross-validation in which the IPC that best
fits the data was found.
}
}
\description{
Returns Bayesian Information Criterion (BIC) information comparing participant data for that ROI to the input IPCs,
in order to determine which IPCs best fit the data. This function is run in a loop by \code{get_data_by_ROI()}.
}
