% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_BIC_at_level.R
\name{run_BIC_at_level}
\alias{run_BIC_at_level}
\title{A function that calculates and stores a \emph{Bayesian Information Criterion (BIC)}
value at the current horizontal and vertical \emph{pattern of interest (POI)}
level.}
\usage{
run_BIC_at_level(
  train_data,
  POIs_list,
  POI_names,
  POIs_to_use,
  analysis_type,
  ROI,
  output_dir,
  held_out_idx,
  CV_log,
  iter
)
}
\arguments{
\item{train_data}{The data on which the \emph{pattern component modelling (PCM)}
model will be trained.}

\item{POIs_list}{A list of \emph{patterns of interest (POI)} which you want
to fit to your data, with the shape \verb{[sample size x number of comparisons]}.}

\item{POI_names}{The names of the \emph{patterns of interest (POI)} which you want
to fit to your data.}

\item{POIs_to_use}{A dataframe containing each previously identified POI for use
in the BIC calculations.}

\item{analysis_type}{Possible values:  \code{'one_sample'}, \code{'two_sample'},
\code{'individual'}, \code{'cross_val'}. A string indicating the type of analysis that
you would like to carry out on the data.}

\item{ROI}{The name of the current \emph{region of interest (ROI)} within which to run the model.}

\item{held_out_idx}{If \code{analysis_type} is \code{cross_val},indicates the held-out values for this iteration.}

\item{iter}{The current iteration (relevant for multi-iteration analyses such as cross-validation and individual analyses).}
}
\description{
A function that calculates and stores a \emph{Bayesian Information Criterion (BIC)}
value at the current horizontal and vertical \emph{pattern of interest (POI)}
level.
}
