% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM_helpers.R
\name{run_ipc_lm}
\alias{run_ipc_lm}
\title{Calculates the Bayesian Information Criterion of a multiple regression analysis
predicting the participant data for a given ROI from the input IPCs.}
\usage{
run_ipc_lm(IPC, ROI_Train, Lv_list)
}
\value{
The Bayesian Information Criterion value (calculated using \code{BIC()}) for the regression, as a float.
}
\description{
Calculates the Bayesian Information Criterion of a multiple regression analysis
predicting the participant data for a given ROI from the input IPCs.
}
