% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PCM.R
\name{run_PCM}
\alias{run_PCM}
\title{A function that accepts input to run a Pattern Component Modelling analysis of Representational Similarity
Analysis (RSA) data.
Accepts input information about the structure of your data.}
\usage{
run_PCM(
  input_dir,
  output_dir,
  POI_file,
  analysis_type,
  holdout = 2,
  num_iters = 1000
)
}
\arguments{
\item{input_dir}{A string to the folder containing an input file with a name defined in \code{POI_file}, as well
as RSA data files in CSV format for each \emph{region of interest (ROI)} (whose names must end with the string \verb{_vert.csv} and
which must be structured in the form \verb{[sample size x number of comparisons]}).}

\item{output_dir}{A string to the folder where you would like all outputs to be written.}

\item{POI_file}{A string defining the path to a CSV file containing one row for each  \emph{pattern of interest (POI)} which you want
to fit to your data, and one column for each comparison made, indicating the pattern of expected correlations for that POI. The first column
of this file should be the names of the POIs.  See (Kryklywy, Ehlers, Beukers, et al., 2020) for a discussion of the setup of -
and rationale behind - POIs.}

\item{analysis_type}{Possible values:  \code{'one_sample'}, \code{'two_sample'}, \code{'individual'}, \code{'cross_val'}. A string indicating the
type of analysis that you would like to carry out on the data.
\itemize{
\item \code{one_sample}: Train a model on one sample and test it on the same sample.
\item \code{two_sample}: Train a model on one sample and test it on a different sample.
\item \code{individual}: Train a model on one sample and test it on one or more individuals' data.
\item \code{cross_val}: Cross-validation - iteratively train a model on one sample and test it on a held-out sample.
}}

\item{holdout}{(default = 2) If \code{analysis_type} is \code{cross_val}, is an integer defining the number of values in your data to hold out on each iteration.}

\item{num_iters}{(default = 1000) If \code{analysis_type} is \code{cross_val}, is an integer defining the number CV iterations to run.}
}
\value{
a summary of PCM data indicating the POIs that best fit the data.
}
\description{
A function that accepts input to run a Pattern Component Modelling analysis of Representational Similarity
Analysis (RSA) data.
Accepts input information about the structure of your data.
}
