% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCM_helpers.R
\name{run_ROI_lm}
\alias{run_ROI_lm}
\title{Conducts a multiple regression analysis on identified scores, predicting the participant
data for a given ROI against the identified IPCs only.}
\usage{
run_ROI_lm(ROI, Lv_list)
}
\value{
An \code{lm()} object (linear regression output).
}
\description{
Conducts a multiple regression analysis on identified scores, predicting the participant
data for a given ROI against the identified IPCs only.
}
