% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{A function that splits data into training and testing data, depending on the analysis type.}
\usage{
split_data(data_for_ROI, analysis_type, holdout)
}
\arguments{
\item{analysis_type}{Possible values:  \code{'one_sample'}, \code{'two_sample'}, \code{'individual'}, \code{'cross_val'}. A string indicating the
type of analysis that you would like to carry out on the data.
\itemize{
\item \code{one_sample}: Train a model on one sample and test it on the same sample.
\item \code{two_sample}: Train a model on one sample and test it on a different sample.
\item \code{individual}: Train a model on one sample and test it on one or more individuals' data.
\item \code{cross_val}: Cross-validation - iteratively train a model on one sample and test it on a held-out sample.
}}

\item{holdout}{(default = 2) If \code{analysis_type} is \code{cross_val}, is an integer defining the number of values in your data to hold out on each iteration.}

\item{input_filename}{A file containing \emph{representational similarity analysis (RSA)} data for a given \emph{region of interest (ROI)},
in the form \verb{[sample size x number of comparisons]}. This file's name should be of the format \code{ROI_vert.csv}, where \code{ROI} is the
name of a brain region from which the data was obtained (e.g. \code{ACC}).}
}
\description{
A function that splits data into training and testing data, depending on the analysis type.
}
