% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_loop.R
\name{train_test_loop}
\alias{train_test_loop}
\title{A function that handles each iteration of training and testing a PCM model.}
\usage{
train_test_loop(
  data_for_ROI,
  ROI,
  POI_names,
  POIs_list,
  analysis_type,
  holdout,
  best_BICs,
  output_dir
)
}
\arguments{
\item{ROI}{The name of the current \emph{region of interest (ROI)} within which to run the model.}

\item{analysis_type}{Possible values:  \code{'one_sample'}, \code{'two_sample'}, \code{'individual'}, \code{'cross_val'}. A string indicating the
type of analysis that you would like to carry out on the data.
\itemize{
\item \code{one_sample}: Train a model on one sample and test it on the same sample.
\item \code{two_sample}: Train a model on one sample and test it on a different sample.
\item \code{individual}: Train a model on one sample and test it on one or more individuals' data.
\item \code{cross_val}: Cross-validation - iteratively train a model on one sample and test it on a held-out sample.
}}

\item{holdout}{(default = 2) If \code{analysis_type} is \code{cross_val}, is an integer defining the number of values in your data to hold out on each iteration.}

\item{input_filename}{A file containing \emph{representational similarity analysis (RSA)} data for a given \emph{region of interest (ROI)},
in the form \verb{[sample size x number of comparisons]}. This file's name should be of the format \code{ROI_vert.csv}, where \code{ROI} is the
name of a brain region from which the data was obtained (e.g. \code{ACC}).}

\item{num_iters}{(default = 1000) If \code{analysis_type} is \code{cross_val}, is an integer defining the number CV iterations to run.}
}
\description{
A function that handles each iteration of training and testing a PCM model.
}
