% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMforR.R
\docType{package}
\name{PCMforR}
\alias{PCMforR}
\title{PCMforR: Pattern Component Modelling for R}
\description{
This package allows you to easily conduct a Pattern
Component Modelling (PCM) analysis on Representational Similarity
Analysis (RSA) data (e.g. from functional brain imaging data) in
R. PCM allows you to take a set of hypothetical structures - Patterns
of Interest, or POIs - by which data might be represented in a given
brain region - and compare these structures to your data to determine which
hypothesis best fits your data using a Bayesian Information Criterion (BIC)
metric. For any additional questions on functionality and use,
please contact: j.kryklywy\link{at}psych.ubc.ca
}
