% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PCM.R
\name{run_PCM}
\alias{run_PCM}
\title{A function that accepts input to run a Pattern Component Modelling analysis of Representational Similarity
Analysis (RSA) data.
Accepts input information about the structure of your data.}
\usage{
run_PCM(
  input_path,
  output_path,
  ROIs,
  IPC_names,
  IPC_file,
  IPC_prior_level,
  DataType,
  Analysis,
  SampleSize,
  num_Conditions,
  Condition_names,
  Hold,
  cv_iterations = 1000
)
}
\arguments{
\item{input_path}{A string to the folder containing an input file with a name defined in \code{IPC_file}, as well
as RSA data files for each ROI (which must be structured in the form \verb{[sample size x number of comparisons]}).}

\item{output_path}{A string to the folder where you would like all outputs to be written.}

\item{ROIs}{A vector of strings naming \emph{regions of interest (ROIs)} in your dataset. PCMs will be run on
data from each ROI in the order named here. Make sure that your RSA data is split by ROI into individual
.csv files.}

\item{IPC_names}{A vector of strings naming \emph{information pattern components (IPCs)} to which you want
to fit your data. Each IPC will be represented as a row in the file defined in \code{IPC_file}, equal in length
to the number of comparisons in your dataset (which is equal to the cumulative sum of the number of
conditions in your dataset). Each cell of this row should contain the expected correlation
of that comparison. See (Kryklywy, Ehlers, Beukers, et al., 2020) for a discussion of the setup of -
and rationale behind - IPCs.}

\item{IPC_file}{A string defining the path to a CSV file containing one row for each IPC (see \code{IPC_names}),
and one column for each comparison made, indicating the pattern of expected correlations for that IPC.}

\item{IPC_prior_level}{A vector containing integers for each level of priors that you would like to check.}

\item{DataType}{A string indicating the type of data structure that you would like to use. Currently,
PCMforR only supports \code{'US'} (Unconditioned stimulus) as a data input.}

\item{Analysis}{Possible values:  \code{'Group'}, \code{'CV'}, \code{'Individual'}, \code{'Mixed'}. A string indicating the
type of analysis that you would like to carry out on the data. \code{'Group'} will analyze data for all
participants as a group. \code{'CV'} will conduct a cross-validation analysis on the group data,re-running the
IPC comparison to the data iteratively in order to achieve the best possible fit on the data.
\code{'Individual'} and \code{'Mixed'} are currently not fully implemented.}

\item{SampleSize}{An integer indicating the number of participants in your data.}

\item{num_Conditions}{An integer indicating the number of trial conditions in your dataset. This will be
used to determine the number of comparisons in your dataset.}

\item{Condition_names}{A vector of strings naming the conditions in your dataset (NOTE: not currently used).}

\item{Hold}{If using cross-validation (CV), is an integer defining the number of values to hold out on each iteration.}

\item{cv_iterations}{If using cross-validation (CV), is an integer defining the number CV iterations to run.}
}
\value{
Does not return a specific output, but outputs a summary of PCM data indicating the IPCs that best fit the data.
}
\description{
A function that accepts input to run a Pattern Component Modelling analysis of Representational Similarity
Analysis (RSA) data.
Accepts input information about the structure of your data.
}
